// packages/payloadset/packages/http/src/Payload/types/MimeTypes.ts
// TODO: Bring in from some library or package for completeness
export type MimeTypes
  = | 'application/json' // JSON format
    | 'application/octet-stream' // Binary data
    | 'application/pdf' // PDF format
    | 'application/xhtml+xml' // XHTML format
    | 'application/xml' // XML format
    | 'image/gif' // GIF image format
    | 'image/jpeg' // JPEG image format
    | 'image/png' // PNG image format
    | 'text/html' // HTML format
    | 'text/plain' // Plain text format
    | 'text/xml' // XML format (alternative)